% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala2e.R
\docType{package}
\name{ala2e}
\alias{ala2e}
\title{Datasets from Fitzmaurice G.M, Laird N.M, and Ware J.H,
"Applied Longitudinal Analysis" (2nd ed., 2011)}
\source{
Fitzmaurice G.M, Laird N.M, and Ware J.H (2011) Applied
        Longitudinal Analysis, 2nd ed., New York: Wiley
}
\description{
This package is an unofficial companion to the textbook
"Applied Longitudinal Analysis"; it includes all the datasets used
in the book, both for easy reproducibility and algorithms
benchmarking purposes.
}
